chore: move state machine and orchestration out of the UI component into a dedicated hook
